+++ CHANGELOG.md
@@ -10,7 +10,8 @@ New:
 Changed:
 - Disable klib signature clash checks for JS compilations. These occasionally occur as a result of Compose compiler behavior, and are safe to disable (the first-party JetBrains Compose Gradle plugin also disables them).
 - Remove `Terminal$Size` and use `IntSize` instead in `Terminal#size` for optimization purposes.
-- Remove Color.Bright* constants. Use Color to create the desired color.
+- Remove `Color.Bright*` constants. Use `Color` function to create the desired color.
+- Replace nullable `Color` and `TextStyle` with `Color.Unspecified` and `TextStyle.Unspecified` respectively. Also make `TextStyle` an inline class.
 
 Fixed:
 - Use CRLF line endings to fix rendering when a terminal is in raw mode.